Add Wara Wara Plaza sound, initial testing for playing sounds
Yay! This took way too many tries!

This initializes the mod (replaces the examplemod -> pretendosmpmusic) files, add our sound, and for now ensure that it can be called during a certain event.

Literally, if I just wanted Wara Wara Plaza, I could just check the instance is on the SMP server, and create coordinate limitations and I'd be fine (and also find a different event for movement.)